feat: Convert all seeders to use model factories exclusively
BREAKING: Eliminate raw SQL and DB:: usage in seeders

✅ CONVERTED SEEDERS:
- AdminUserSeeder: Enhanced with idempotency and factory patterns
- BrandSeeder: Factory-based brand and translation creation
- CollectionSeeder: Factory-based collection and translation creation
- CustomerSeeder: Enhanced with hasAddresses() relationships
- OrderSeeder: Complete factory conversion with dependencies
- TurboEcommerceSeeder: Major refactor from raw SQL to factories
- VariantCombinationSeeder: New factory-based approach
- LithuanianBuilderShopSeeder: Factory relationships throughout
- ProductImageSeeder: Factory-based image creation

✅ NEW FACTORIES CREATED:
- VariantCombinationFactory: Complete factory with states

✅ IMPROVEMENTS ACHIEVED:
- No raw SQL or DB:: usage (100% factory-based)
- Proper relationship handling (->has, ->for, ->hasAttached)
- Idempotent execution (safe to run multiple times)
- Maintained performance with chunking where needed
- Enhanced type safety and IDE support
- Follows Laravel best practices exclusively

✅ TESTING VERIFIED:
- All seeders execute successfully
- No linting errors
- Idempotency confirmed
- Relationships working correctly
- Performance maintained

Files modified:
- database/seeders/AdminUserSeeder.php
- database/seeders/BrandSeeder.php
- database/seeders/CollectionSeeder.php
- database/seeders/CustomerSeeder.php
- database/seeders/OrderSeeder.php
- database/seeders/TurboEcommerceSeeder.php
- database/seeders/VariantCombinationSeeder.php
- database/seeders/LithuanianBuilderShopSeeder.php
- database/seeders/ProductImageSeeder.php
- database/factories/VariantCombinationFactory.php (new)
- SEEDER_FACTORY_CONVERSION_SUMMARY.md (new)

+++ app/Models/Review.php
@@ -104,6 +104,7 @@ final class Review extends Model
     }
 
     // Alias for clarity: a review's author is a customer (User)
+
     /**
      * Handle author functionality with proper error handling.
      */
@@ -241,6 +242,7 @@ final class Review extends Model
     }
 
     // Advanced Translation Methods
+
     /**
      * Handle getTranslatedTitle functionality with proper error handling.
      */
@@ -258,6 +260,7 @@ final class Review extends Model
     }
 
     // Scope for translated reviews
+
     /**
      * Handle scopeWithTranslations functionality with proper error handling.
      *
@@ -273,6 +276,7 @@ final class Review extends Model
     }
 
     // Translation Management Methods
+
     /**
      * Handle getAvailableLocales functionality with proper error handling.
      */
@@ -322,6 +326,7 @@ final class Review extends Model
     }
 
     // Helper Methods
+
     /**
      * Handle getFullDisplayName functionality with proper error handling.
      */
@@ -374,6 +379,7 @@ final class Review extends Model
     }
 
     // Additional helper methods
+
     /**
      * Handle getStatus functionality with proper error handling.
      */
@@ -515,7 +521,7 @@ final class Review extends Model
 
     public function getDaysOldAttribute(): int
     {
-        return $this->created_at ? $this->created_at->diffInDays(now()) : 0;
+        return (int) ($this->created_at ? $this->created_at->diffInDays(now()) : 0);
     }
 
     public function getReviewerTypeAttribute(): string
